requirements.txt
asgiref==3.8.1
Django==5.2
django-debug-toolbar==5.2.0
django-filter==25.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-nested-forms==1.1.8
drf-spectacular==0.28.0
Faker==37.1.0
pillow==11.2.1
PyJWT==2.9.0
sqlparse==0.5.3
tzdata==2025.2
psycopg2-binary==2.9.9